requirements.txt
absl-py==1.0.0
astunparse==1.6.3
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
chex==0.1.0
cycler==0.11.0
dm-tree==0.1.6
flatbuffers==2.0
flax==0.3.6
fonttools==4.33.3
gast==0.3.3
gin-config==0.5.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.41.1
h5py==2.10.0
idna==3.3
imageio==2.13.0
imageio-ffmpeg==0.4.5
jax==0.2.22
jaxlib==0.1.72
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
Markdown==3.3.4
matplotlib==3.5.2
msgpack==1.0.2
numpy==1.18.5
oauthlib==3.1.1
opencv-python==4.5.4.58
OpenEXR==1.3.2
opt-einsum==3.3.0
optax==0.1.0
packaging==21.3
Pillow==8.4.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyMCubes==0.1.2
pyparsing==3.0.6
python-dateutil==2.8.2
PyYAML==5.3.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
Rtree==0.9.7
scipy==1.7.2
six==1.16.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
toolz==0.11.2
tqdm==4.62.3
trimesh==3.9.35
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
